Allow paths such as ":one/:b" and ":two/aaa" to coexist.
Ambiguous wildcards such as "one/:a/two" and "one/:b/two" are still disallowed.
Fixes #13

With this method, each handler node stores an array of wildcard names that
lead up to it. Once the proper node is found, as the recursion unwinds
the list of wildcards seen in the request is built, and this list,
along with the array of wildcard names, forms the parameter map which
is passed to the handler.

This also converts wildcard nodes from a list to a single node representing
all possible wildcards for that portion of the tree.