Match less specific routes when appropriate.
When a node is found but it does not handle the requested HTTP method,
look for a less specific match in the rest of the tree that does handle
that method.

This also changes in the behavior of HeadCanUseGet, which is now
checked at route add time instead of at request time. This means that
if you are setting HeadCanUseGet to false, you must do so before
adding your routes.

Fixes #27